writer/requirements.txt
paho-mqtt
psycopg2-binary
pandas
prophet